Remov the window state setting for DisplayWidget
For a DisplayWidget that is called from user scripts, you do not need to set the window size, since its size depends on the parent widget and the layout in which we put it. Moreover, for Qt 5.15, because of this line, the widget DisplayWidget is not displayed at all, since it has already been installed in layout.